+++ examples/addons/pycsg_sphere.py
@@ -8,6 +8,7 @@ from ezdxf.render.forms import sphere, cube
 from ezdxf.addons.pycsg import CSG
 
 DIR = Path('~/Desktop/Outbox').expanduser()
+NLENGTH = .05
 
 doc = ezdxf.new()
 doc.layers.new('csg', dxfattribs={'color': 1})
@@ -21,14 +22,14 @@ sphere1 = sphere(count=32, stacks=16, radius=.5, quads=True)
 
 union = (CSG(cube1) + CSG(sphere1)).mesh()
 union.render(msp, dxfattribs={'layer': 'csg', 'color': 1})
-union.render_normals(msp, dxfattribs={'layer': 'normals'})
+union.render_normals(msp, length=NLENGTH, relative=False, dxfattribs={'layer': 'normals'})
 
 subtract = (CSG(cube1) - CSG(sphere1)).mesh().translate(2.5)
 subtract.render(msp, dxfattribs={'layer': 'csg', 'color': 3})
-subtract.render_normals(msp, dxfattribs={'layer': 'normals'})
+subtract.render_normals(msp, length=NLENGTH, relative=False, dxfattribs={'layer': 'normals'})
 
 intersection = (CSG(cube1) * CSG(sphere1)).mesh().translate(4)
 intersection.render(msp, dxfattribs={'layer': 'csg', 'color': 5})
-intersection.render_normals(msp, dxfattribs={'layer': 'normals'})
+intersection.render_normals(msp, length=NLENGTH, relative=False, dxfattribs={'layer': 'normals'})
 
 doc.saveas(DIR / 'csg_sphere.dxf')