+++ NEWS.md
@@ -53,7 +53,8 @@ Version 0.11b3 - dev
 - NEW: `ezdxf.math.intersection_ray_ray_3d()` calculate intersection for 3D rays 
 - NEW: `ezdxf.math.Plane()` 3D plane construction tool 
 - NEW: `ezdxf.render.MeshTransformer()` simple inplace mesh transformations class
-- NEW: `MeshBuilder.render_normals` render face normals as LINE entities, useful to check face orientation
+- NEW: `MeshBuilder.render()` added UCS support
+- NEW: `MeshBuilder.render_normals()` render face normals as LINE entities, useful to check face orientation
 - NEW: `ezdxf.render.forms.cone_2p()` create 3D cone mesh from two points
 - NEW: `ezdxf.render.forms.cylinder_2p()` create 3D cylinder mesh from two points
 - NEW: `ezdxf.render.forms.sphere()` create 3D sphere mesh